// src/api/models/notice-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sanchalak Mobile API
 * API documentation for Sanchalak Mobile Application (Parent/Student)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@sanchalak.in
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface NoticeDto {
    'id'?: number;
    'title'?: string;
    'priority'?: string;
    'targetRole'?: string;
    'publishDate'?: string;
    'expiryDate'?: string;
    'attachmentUrl'?: string;
    'createdByName'?: string;
    'read'?: boolean;
}